requirements.txt
altgraph==0.12
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
ipython==4.2.0
ipython-genutils==0.1.0
macholib==1.7
modulegraph==0.12.1
pbr==1.10.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
PyInstaller==3.2
PyYAML==3.11
selenium==2.53.6
simplegeneric==0.8.1
six==1.10.0
stevedore==1.15.0
traitlets==4.2.1
virtualenv==15.0.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
